requirements.txt
google-generativeai
pandas
matplotlib
seaborn
scikit-learn
statsmodels
numpy
joblib
streamlit